980 fix chartList
